-observations splines

+++ src/constants.f90
@@ -6,7 +6,7 @@ MODULE constants
 
   INTEGER, PARAMETER :: mnsp=250, mre=2000
   INTEGER  :: i
-  real(kind=4):: DFRACT
+  real(dp):: DFRACT
   ! variables for zenith routine which calculates zenith angle
   REAL(dp) :: theta, secx, cosx
 ! generic reaction rate variables
@@ -52,6 +52,9 @@ CONTAINS
 
 ! Time2 is local time in hours
     Time2=mod(Time/(60.*60.), 24.)
+   
+  
+
     IF (TIME2 .LT. 0) TIME2=TIME2+24.
     LAT2=LAT
 
@@ -128,6 +131,8 @@ CONTAINS
     DO i=1,jmax
       IF (j(i) .lt. 0.e0) j(i)=0.d0
     ENDDO
+
+
 
 
   END SUBROUTINE mcm_constants
